refactor(cap-service): refactored cap-service-folder
