+++ src/cap-service/index.ts
@@ -1,13 +1,13 @@
-export * from './commitlint';
-export * from './devcontainer';
-export * from './eslint';
-export * from './jest';
-export * from './git';
-export * from './github';
-export * from './husky';
-export * from './npm';
+export * from './config/commitlint';
+export * from './config/devcontainer';
+export * from './config/eslint';
+export * from './config/jest';
+export * from './config/git';
+export * from './config/github';
+export * from './config/husky';
+export * from './config/npm';
 export * from './project';
-export * from './prettier';
-export * from './vscode';
+export * from './config/prettier';
+export * from './config/vscode';
 export * from './typescript';
-export * from './samplecode';
+export * from './config/samplecode';
